Fix URL for metar and add taf command
